+++ tests/unit/scala/test_utils_suite.py
@@ -444,7 +444,7 @@ def test_is_sql_select_statement():
             with anon_sproc as procedure
             ' as col1
            ) select col1 from t
-        """
+        """,
     ]
     for s in select_sqls:
         assert is_sql_select_statement(s)