+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - `array_max`
 - Added support for async execution of multi-query dataframe containing binding variables.
 - Added support for renaming multiple columns in `DataFrame.rename`.
+- Added support for Geometry datatypes.
 
 ### Improvements